--- projeto-crm-franquias/backend/src/main/java/com/crm/franquias/model/Academia.java
@@ -1,18 +1,0 @@
-package com.crm.franquias.model;
-
-public class Academia extends Franquia {
-    private String checkin;
-
-    public Academia(int id, String nome, String cidade, String status, String checkin) {
-        super(id, nome, cidade, status);
-        this.checkin = checkin;
-    }
-
-    public String getCheckin() {
-        return this.checkin;
-    }
-
-    public void setCheckin(String checkin) {
-        this.checkin = checkin;
-    }
-}
